+++ Intent.Modules.Java.Persistence.JPA/release-notes.md
@@ -3,6 +3,7 @@
 - Names of generated CompositeIds are now always PascalCased.
 - Fixed: `@Table` annotation wasn't being added for derived types.
 - Regard attributes named `id` implicitly as being primary keys when no explicit primary key.
+- Fixed: `cascade = javax.persistence.CascadeType.ALL` was being indiscriminately applied to all relationships, it will now only be applied on navigations from a composite owner.
 
 ### Version 4.0.0